// pronti-frontend-main/front/src/app/modulos/usuario/form/form-usuario.enum.ts
export enum FormUsuarioEnum {
  // nombre = 'nombre',
  username = 'username',
  password = 'password',
  nombres = 'nombres',
  apellidos = 'apellidos',
  fechaNacimiento = 'fechaNacimiento',
  correo = 'correo',
  rolUsuario = 'rolUsuarioCollection',

  tipoMedioContacto = 'tipoMedioContacto1',
  medioContacto = 'medioContacto1',
  tipoDocumentoIdentidad = 'tipoDocumentoIdentidad',
  documentoIdentidad = 'documentoIdentidad',
  pais = 'pais',
  provincia = 'provincia',
  ciudad = 'ciudad',
  direccion = 'direccion'
}